+++ Common/Interfaces/IVehicle.cs
@@ -7,8 +7,8 @@ public interface IVehicle
     string PlateNumber { get; init; }
     string Make {  get; init; }
     int Odometer { get; init; }
-    double CostPerKm { get; init; }
-    double CostPerDay { get; init; }
+    int CostPerKm { get; init; }
+    int CostPerDay { get; init; }
     VehicleTypes VehicleType { get; init; }
     VehicleStatuses VehicleStatus { get; init; }
 }
